requirements.txt
fastapi
uvicorn
simple-salesforce
pandas
python-dotenv
openai
jinja2
python-multipart
